feat: notification List 반환 로직 이관
