+++ src/main/java/com/example/diary/domain/member/service/MemberService.java
@@ -3,8 +3,6 @@ package com.example.diary.domain.member.service;
 import com.example.diary.domain.member.dto.InfoResponseDto;
 import com.example.diary.domain.member.dto.MyInfoRequestDto;
 import com.example.diary.domain.member.entity.Member;
-import com.example.diary.domain.notification.dto.NotificationPagingDto;
-import org.springframework.data.domain.Pageable;
 
 public interface MemberService {
 
@@ -19,5 +17,5 @@ public interface MemberService {
 
     void withdrawMembership(Member member);
 
-    NotificationPagingDto notificationList(Member member, Pageable pageable);
+
 }
